kuhn game skeleton

+++ src/player/cfr_player.rs
@@ -2,11 +2,11 @@ use crate::kuhn;
 use crate::player::player;
 
 struct CfrPlayer {
-    player_id: i64,
+    player_id: i32,
 }
 
 impl player::Player for CfrPlayer {
-    fn on_register(&mut self, player_id: i64) {
+    fn on_register(&mut self, player_id: i32) {
         self.player_id = player_id;
     }
     fn decide_action(&mut self, game_info: &kuhn::ActionHistory) -> kuhn::Action {
