+++ src/player/player.rs
@@ -1,7 +1,7 @@
 use crate::kuhn;
 
 pub trait Player {
-    fn on_register(&mut self, player_id: i64);
+    fn on_register(&mut self, player_id: i32);
     fn decide_action(&mut self, game_info: &kuhn::ActionHistory) -> kuhn::Action;
     fn handle_result(&mut self, game_info: &kuhn::ActionHistory, payoff: i64);
 }
